requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.13.0

# Elasticsearch
elasticsearch>=8.11.0

# Kafka
confluent-kafka>=2.3.0

# Redis
redis>=5.0.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# JWT & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Document parsing
pymupdf>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
unstructured>=0.11.0
python-multipart>=0.0.6

# OCR - PaddleOCR-VL
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0

# Multimodal (Image & Video)
opencv-python>=4.8.0
pillow>=10.0.0

# Vector embedding
# 使用阿里云qianwen3-embedding API

# LLM clients
openai>=1.3.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0
pyyaml>=6.0.0

# InfluxDB
influxdb-client>=1.38.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
